M4: Split FileUploadInput mega-test - Extract error handling concerns
Reduced main test file from 804 lines to 602 lines (202 line reduction).
Created error-handling-and-deletion.unit.test.tsx with 227 lines (6 tests).

Extracted test group:
- error handling and deletion (6 tests)

All 1421 tests passing. Test file count now 121.
